tds-project-1/requirements.txt
# Existing
fastapi>=0.109.0
requests>=2.28.0,<3
python-dotenv>=1.0.0
uvicorn>=0.27.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
duckdb>=0.9.0
numpy>=1.24.0
python-dateutil>=2.8.0
docstring-parser>=0.15.0
httpx>=0.26.0
pydantic>=2.0.0
pytesseract>=0.3.10
pillow>=10.0.0
ffmpeg-python>=0.2.0
pydub>=0.25.0
SpeechRecognition>=3.10.0

# LangChain, LangSmith, LiteLLM, Pydantic AI, GPTCache
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-classic>=0.3.0
langsmith>=0.1.0
langchain-litellm>=0.0.1
litellm>=1.30.0
pydantic-ai>=0.0.14
gptcache>=0.1.0
openai>=1.0.0
